mlops-multi-account-cdk/mlops-infra/requirements.txt
awscliv2==2.2.0
boto3==1.28.63
aws-cdk-lib==2.100.0
dataclasses-json==0.6.1
constructs>=10.0.0,<11.0.0
yamldataclassconfig==1.5.0
aws_cdk.aws_lambda_python_alpha==2.100.0a0
